requirements.txt
ibm-watson-machine-learning
python-dotenv
PyPDF2
python-docx
pytesseract
PyMuPDF
pdf2image
openpyxl
